correction bug correspondance type

+++ diabet/src/diabet/Donnees.java
@@ -60,7 +60,7 @@ public class Donnees {
 		return taille;
 	}
 
-	public void setTaille(float taille) {
+	public void setTaille(double taille) {
 		this.taille = taille;
 	}
 
@@ -68,7 +68,7 @@ public class Donnees {
 		return poids;
 	}
 
-	public void setPoids(float poids) {
+	public void setPoids(double poids) {
 		this.poids = poids;
 	}
 
@@ -84,7 +84,7 @@ public class Donnees {
 		return tourDeTaille;
 	}
 
-	public void setTourDeTaille(float tourDeTaille) {
+	public void setTourDeTaille(double tourDeTaille) {
 		this.tourDeTaille = tourDeTaille;
 	}
